post.delete complete

+++ back/controllers/post.controller.js
@@ -49,12 +49,24 @@ exports.modifyPost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   Post.findOne({ _id: req.params.id })
     .then(post => {
-      const filename = post.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Post.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Post supprimé !'}))
-          .catch(error => res.status(400).json({ error }));
-      });
+      if (post.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        if (post.imageUrl) {
+          const filename = post.imageUrl.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Post.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: 'Post supprimé !' }))
+              .catch(error => res.status(400).json({ error }));
+          });
+        } else {
+
+          Post.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Post supprimé !' }))
+            .catch(error => res.status(400).json({ error }));
+        }
+      }
+
     })
     .catch(error => res.status(500).json({ error }));
 };
